requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.1.18
chardet==3.0.4
idna==2.6
psycopg2-binary==2.7.4
requests==2.18.4
urllib3==1.22
